Improve create package
- Move metafields config to the top level
- Allow merging shopifile.yml config
- Remove unused code for skipping empty directories
- Fix bug in create where it was unnecessarily merging template data for irrelevant template folders
Previously we had a MergeTemplates task that merge data for the global template (package.json.tpl and shopifile.yml.tpl) as well as merge data for a specific extension template. The global files are at the root so this task was also merging data for irrelevant templates.  The fix is to break this into 2 different tasks, 1 to merge global templates and another to merge templates specific to the extension