New template engine (#270)
This PR introduces an entirely new, much more generic render engine. Templates are now spilt into projects/ and shared/. The subdirectories in projects/ are blueprints for the exact structure that is being generated when someone runs the create command. src/index.tpl is the only file that gets minor special treatment as it will either be translated to src/index.{js,ts,tsx} depending on whether React and / or TypeScript were chosen as template settings.

The shared/ directory structure is technically arbitrary but currently mimics the projects/ structure to a large degree. Files in shared/ can be thought of as partials that are either included with {{ file "shared/... }} in their raw form or interpreted with the template engine via {{ template "shared/..." . }}. The . is critical when running template as it signals the template engine to pass the entire data object it received during invocation to the partial.

This PR also introduce a new, generic merge function (curtesy of @dmhenry) that can be used in templates to deep merge YAML or JSON files.