[Feature] Add logic to install all app/home dependencies (#195)
* Slightly leaner way to get the branch from the URL

* Add the real branch name to the init command.  This way we can work with a real template

* Instead of using hooks automatically install npm dependencies

To do this we:

1. Remove the hook functionality from create-app.  create-app no longer searches for, or executes hooks
2. In the 'installing dependencies' step we use a glob file path to find every package.json file
3. We remove package.json from these file paths to get the path to the parent folder.
4. For every folder we run a task (up to 3 concurantly) to install those dependencies
5. We track the number of tasks that have completed.  When all the dependencies are installed we update the task title.

* Remove cleanup of files that will no longer apply if we remove hooks

The _templates folder is only used to store hooks.  In this branch we assume that hooks are being replaced with automatically installing dependencies.  If this happens we'll remove the _template folder from the templates. Therefore we won't need to remove them

* Added tests for the logic to install dependencies

The approach here is to abstract the logic into a utility, and then test that utility.  This leaves the execution of that utility within services/init untested.  I think this would be better suited for an integration test.  I'm going to leave this for now until I get a run down on how integration tests work from Pedro

* Fix type error by explicitly typing dependency.install

Previosuly TS didn't realise this is a mock

* PR Feedback:

1. Add some type safety to the function for recursively installing dependencies
2. Rename that function so it's a little clearer it's NPM dependencies, and that it's a deep installation

* Change how we construct the titlePath

The previous implementation relied on 2 things:

1. It being a unix file system that uses / as path separators
2. There being an /app/ folder at the end of the path

The new implementation doesn't rely on either of these.