Redirect /dev-console/assets to the embedded file system nested path
The embedded Dev Console serve all of its assets from a relative path but the embedded
file system is served from "/dev-console" path. We need to add a custom handler so when the path is "/dev-console/assets" we redirect to "/dev-console/dev-console/assets".

The nesting is ugly but necessary to get our reverse proxy working for serving dev console assets under
single path.