Sort apps by name when presented with prompt
This PR proposes to sort the list of apps when presented with the
app select prompt. This is useful when you have a lot of apps and
you want to find a specific one quickly. It was one of the pieces
of feedback that we heard from folks at Unite.

Co-Authored-By: David Witt <david.witt@shopify.com>
Signed-off-by: Steve Martinelli <steve.martinelli@shopify.com>