Change how the app loads entrySourceFilePaths
Previously we did a glob search to find every /src/*.{js|jsx|ts|tsx} file.  When I looked at the unit tests I realized 2 things:

1. Currently the CLI supports files in /*.{js|jsx|ts|tsx}
2. For the new ui_extension type we should be getting the src files from shopify.ui.extension.toml because that is the source of truth.

As such I needed to rework this. We now do a check on the shopify.ui.extension.toml schema.

* If the schema matches the old schema we use the old code for getting entrySourceFilePaths.
* If the schema macthes the old schema we compile the fullpaths extensionPoints array in the shopify.ui.extension.toml file.

This branching logic is kinda complex, and a bit of a shame, but it should go away aftr Isaac's refactor, and/or after we migrate every UI extension to the new format.