Updated the URL to link to the embeeded app
Context: When the user installs the Shopify app (android/ios) the app registers some URL's with the operating system.  This tells the OS 'Hey, I handle URL's of this shape'.  Then when the user scans the QR Code, the OS says 'hey Shopify handles this URL' and opens the Shopify app.  From there the Shopify app redirects to the embedded app view. Amazing