Add command guidelines to the contributor docs
To ensure consistency and quality in the Shopify CLI, we have guidelines around commands. We're adding those guidelines into our contributor documentation to scale our teams' work to other contributing teams.

This is a V1. Future work on this doc can include a list of cross-domain flags, more guidance around how env variables work as flags, and more examples gleaned from decisions we've made (for example, we could draw on work on webhooks or the docs search feature).