Watch all non-hidden GrahpQL files in functions
Because we don't know how people will name their GraphQL files, it's
best not to force a `input.*` prefix. The main reason for this prefix
was because our current Rust crate touches a hidden GraphQL file which
sent the CLI in a build loop (rebuild triggers a file change, which
triggers a rebuild, rinse & repeat).

Instead of using a prefix, this commit fixes that by globbing on all
non-hidden graphql file.