Update README.md
Still needs some work, but updating to support the upcoming `npx` option as a unified recommended path to scaffold. 

I want to push this further and make each of the three options listed here look and feel the same.

So far I've added in the new instructions, and updated some broken links to the docs.