quick and dirty fix for lost binding
When `developerPlatformClient.refreshToken` is passed to
`performActionWithRetryAfterRecovery` it is invoked standalone without
the instance as context and so the `this` value defaults to the global
and errors with `TypeError: this.session is not a function`.

We could use class fields in `AppManagementClient` to permanently bind
bind the methods with a closure at their creation, so we don't have to
bind them ad hoc. But for now it's just quick and dirty to directly
resolve the issue Phong and I ran into.