Use correct protocol based on browser
Even if the browser connects via an HTTPS tunnel, the server only sees
http in req.protocol because it exposes http on localhost.

So we also have to look at the X-Forwarded-Proto header, which should
indicate that the original request was HTTPS. This will allow us to send
back the correct URL for the ping, status, and graphql endpoints.