Add vendored DevServer library (#4825)
### WHY are these changes introduced?

Internal to Shopify, the CLI needs to support the new dev server infrastructure and hostnames for local development environments.

For any non-Shopifolk: there's no difference, it's an internal adjustment.

### WHAT is this pull request doing?

Adds a copy of the DevServer library (with permission) and use it for FQDNs during local dev.

### How to test your changes?

Run ` SHOPIFY_SERVICE_ENV="local" p shopify app info --path=<path to an app> --verbose`

You'll see the CLI exit, because the first thing it wanted to connect locally to wasn't there (probably `identity`). Confirm the error message and the URL it was trying to connect to.

### Checklist

- [x] I've considered possible cross-platform impacts (Mac, Linux, Windows)
- [x] I've considered possible documentation changes

### Measuring impact

How do we know this change was effective? Please choose one:

- [x] n/a - this doesn't need measurement, e.g. a linting rule or a bug-fix