Make types a little more explicit
Previously the options passed to generateFrontendURL were an object with a bunch of optional properties.  But if noTunnelUseLocalhost is true then port is mandatory and it's alwasy passed.  By switching options to a union typescriot understands this and we no longer need to have get an available TCP port for localhost... it would alwasy be passed