Update path flag language to be more generic
While for almost all commands a theme folder is required that's not
necessarily true for some of them. Instead of modifying the flag
description for some of the commands we can generalize the description
in a way that explains exactly what it's doing while covering both
scenarios.

Users will get a warning that the command is not being run in a theme
directory anyways.