Refactor --unpublished to accept a value of create or upsert
In CI contexts, there are times when we want to create a theme that
matches a branch name. For this use case, we want our use of unpublished
to be idempotent so subsequent pushes update theme with the name that
matches the branch.

After considering a couple options, it was decided that the best option would be
to change --unpublished from a boolean flag to a string flag with a default.
This will preserve the existing behaviour and given an escape hatch for those
who want the idempotent behaviour.