Warn if path is set via CLI during multi env command runs
The path flag should point to where a command
will be run. This defaults to the current working
directory

While running a command with multiple environment
flags, providing a path via the CLI will take
precedence over toml and default values and
try to run every environment at the same path. This
PR prevents this by displaying a warning message