Set projectRoot to null explicitly
This is important because if we don't send it explicitly and just
leave it out, it defaults to the current working directory.

Generally speaking, this will result in all stack frames from the
CLI being identified as outside of the project, preventing correct
error grouping in Observe.

By setting an explicit null project root, we prevent the SDK-side
assignment of in-project state, and we fall back to Observe's
much smarter algorithm for identifying in-project lines.

For context, see https://github.com/bugsnag/bugsnag-js/blob/286caed5daae9349be8fb22ab0051204bfc1a813/packages/plugin-node-in-project/in-project.js