Fix images not loading in theme app extensions
Our wrapper around fs's `readFile` sets a default of `{encoding:
'utf8'}` to the options. This will cause any binary files (like images
or fonts) to not be read properly.

There isn't a way to disable the encoding based on how the function is
currently typed so I'm switching to fs's version of `readFile` which
will handle all file types correctly automatically.